DockerfileLintBear: Move to configfiles package
