+++ bears/tests/configfiles/DockerfileLintBearTest.py
@@ -6,7 +6,7 @@ from queue import Queue
 sys.path.insert(0, ".")
 import unittest
 from bears.tests.LocalBearTestHelper import LocalBearTestHelper
-from bears.linters.DockerfileLintBear import DockerfileLintBear
+from bears.configfiles.DockerfileLintBear import DockerfileLintBear
 from coalib.settings.Section import Section
